move: Racing unassigned card
